+++ src/Casts/AsTagValue.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace DiscoveryUkraine\SagaLaraFlow\Casts;
+
+use Illuminate\Contracts\Database\Eloquent\CastsInboundAttributes;
+use Illuminate\Database\Eloquent\Model;
+
+class AsTagValue implements CastsInboundAttributes
+{
+    /**
+     * Prepare the given value for storage.
+     *
+     * @param  array<string, mixed>  $attributes
+     */
+    public function set(
+        Model $model,
+        string $key,
+        mixed $value,
+        array $attributes,
+    ): ?string {
+        return $value === null ? null : (string) $value;
+    }
+}